requirements.txt
flask
python-dotenv
pdfplumber
requests
jsonschema
gunicorn